use ProcessConfiguration for outputs and zone id
