+++ cse-runner-app/src/main/resources/application.yml
@@ -35,6 +35,10 @@ minio-adapter:
 cse-cc-runner:
   zone-id: "Europe/Paris"
   trm: 500
+  outputs:
+    initial-cgm: INITIAL_CGM
+    final-cgm: FINAL_CGM
+    ttc-res: TTC_RES
   xnodes:
     file-path: /home/farao/xnodes/xnodes.xml
   combined-ras:
@@ -69,7 +73,3 @@ cse-cc-runner:
   whitelist:
     - http://minio:9000/
     - http://localhost:9000/
-outputs:
-  initial-cgm: INITIAL_CGM
-  final-cgm: FINAL_CGM
-  ttc-res: TTC_RES
